'添加七牛云工具'

+++ myblog/src/main/java/cn/nuofankj/myblog/dto/impl/FriendsTypeDto.java
@@ -11,9 +11,9 @@ import java.util.List;
 
 @Data
 public class FriendsTypeDto {
-    private int id;
+    private long id;
     private String name;
-    private int count;
+    private long count;
     private List<FriendDto> list = new ArrayList<>();
 
     public static FriendsTypeDto toDTO(FriendsType friendsType, List<Friends> list) {
